requirements.txt
paho-mqtt>=1.6.1
RPi.GPIO>=0.7.1
